requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.1
Django==4.2.7
django-ckeditor==6.7.0
django-crispy-forms==2.1
django-js-asset==2.1.0
idna==3.4
Pillow==10.1.0
razorpay==1.4.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
